projects: ad469x_evb: rename ad469x_fmcz
Because the hdl project was renamed to ad469x_evb no hardware file with
the previous naming was found.

To preserve consistency no-os project is also renamed.
