+++ doc/sphinx/source/projects/ad469x_evb.rst
@@ -1,0 +1,1 @@
+.. include:: ../../../../projects/ad469x_evb/README.rst
