--- doc/sphinx/source/projects/ad469x_fmcz.rst
@@ -1,1 +1,0 @@
-.. include:: ../../../../projects/ad469x_fmcz/README.rst
