+++ projects/ad469x_evb/src/common/common_data.c
@@ -1,6 +1,6 @@
 /***************************************************************************//**
  *   @file   common_data.c
- *   @brief  Defines common data to be used by ad469x_fmcz examples.
+ *   @brief  Defines common data to be used by ad469x_evb examples.
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
