+++ projects/ad469x_evb/src/examples/examples_src.mk
