+++ projects/ad469x_evb/src/examples/iio_example/iio_example.h
@@ -1,6 +1,6 @@
 /***************************************************************************//**
- *   @file   main.c
- *   @brief  Main file for xilinx platform of ad469x_fmcz project.
+ *   @file   iio_example.h
+ *   @brief  IIO example header for ad469x_evb project
  *   @author Axel Haslam (ahaslam@baylibre.com)
 ********************************************************************************
  * Copyright 2024(c) Analog Devices, Inc.
@@ -30,38 +30,12 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
+#ifndef __IIO_EXAMPLE_H__
+#define __IIO_EXAMPLE_H__
 
 /******************************************************************************/
-/***************************** Include Files **********************************/
+/************************ Functions Declarations ******************************/
 /******************************************************************************/
-#include "platform_includes.h"
-#include "common_data.h"
-#include "no_os_error.h"
-#ifdef IIO_EXAMPLE
-#include "iio_example.h"
-#elif defined(BASIC_EXAMPLE)
-#include "basic_example.h"
-#endif
+int iio_example_main();
 
-/***************************************************************************//**
- * @brief Main function execution for xilinx platform.
- *
- * @return ret - Result of the enabled examples execution.
-*******************************************************************************/
-int main()
-{
-	int ret = -EINVAL;
-
-	Xil_DCacheDisable();
-	Xil_ICacheDisable();
-
-#ifdef IIO_EXAMPLE
-	ret = iio_example_main();
-#elif defined (BASIC_EXAMPLE)
-	ret = basic_example_main();
-#else
-#error At least one example has to be selected using y value in Makefile.
-#endif
-
-	return ret;
-}
+#endif /* __IIO_EXAMPLE_H__ */
